ASoC: SOF: Intel-HDA - get link dma channel for fw
Host dma and link dma work in decouple mode for SOF+HDA
and they are set independently. The host dma is set by
dai frontend and link dma is set by dai backend. There
is a group of stream registers in HDA controller for each
dma channel. So to make host driver set correct register,
the dma channel is allocated in host and sent to FW for
allocation and host set the register according to this
dma channel.

Signed-off-by: Rander Wang <rander.wang@linux.intel.com>