ASoC: SOF: intel: hda-ipc: fix ipc irq_status modified by stream interrupt issue
We are using share irq for IPC and stream IOC interrupts, we should
check and make sure the interrupt is for IPC before we should store it
to sdev->irq_status, otherwise, the previously stored one may be
modified(to 0 usually) when a consequent IOC interrupt arrives.

Here add check to fix that.

Signed-off-by: Keyon Jie <yang.jie@linux.intel.com>