ASoC: SOF: Intel: hda-ipc: remove irq_status from sdev and fix missing handling to IPC IRQ in the thread
We are using a shared IRQ for IPC and stream IOC interrupts. A race
condition can happen if the first part of the next interrupt deals
with stream events while the interrupt thread of the current interrupt
dealing with IPC hasn't been scheduled yet.

To fix that, We need to check the interrupt source before wakeing up the
handler thread to make sure we handle IPC interrupt here only.

On the other hand, once we are in the thread, checking sdev->irq_status
is superfluous, so here remove irq_status from definition of struct
snd_sof_dev as nobody need it anymore.

Signed-off-by: Keyon Jie <yang.jie@linux.intel.com>