ASoC: SOF: PM: check jack status to detected headphone when resuming
When checking sound setting in ubuntu, hda controller is activated
from suspended mode but codec is still in suspended mode. It is
impossible to detect headphone pluging event for this case. And
I also checked this case with legacy hda driver which also can't
handle this case.

Now the solution is to poll jack status in resuming function.
This operation will activate codec then unsolicted event would work.

Signed-off-by: Rander Wang <rander.wang@linux.intel.com>