Revert "ASoC: SOF: hda: set autosuspend delay for hda bus device"
This reverts commit 69964d0992a29b95b63e519f2dd7ded2e953fd88

This patch causes a bug: playback and catpure don't work
if playback and catpure are started when ubuntu is just set up.
The hda codec is suspended immediately after initialized in
legacy hda driver and then it works after resume. This patch
breaks the flow in SOF, so just remove it.

Signed-off-by: Rander Wang <rander.wang@linux.intel.com>