soundwire: intel: refine function for wakeen event processing
when processing wakeen event, we need to wake up the slaves may
generate this event. This PR refines the processing function to
discard the devices that are defined in ACPI tables but not
physically present and devices that cannot generate wakes.

if a slave has been attached to a bus, the slave->dev_num_sticky
should be non-zero, so we can check this value to skip the
ghost devices defined in ACPI table but not populated in hardware.

Signed-off-by: Rander Wang <rander.wang@intel.com>